Remove Settings page
