+++ src/Mindbank/Views/MainView.axaml
@@ -191,9 +191,45 @@
                     </Button>
                     <Button HotKey="Ctrl+S" HorizontalAlignment="Stretch" HorizontalContentAlignment="Center"
                             ToolTip.Tip="{x:Static lang:Lang.ToolTip_Settings}" Background="Transparent"
-                            Tag="{CompiledBinding #SettingsView}" Click="NoteGroupButtonClicked"
                             DockPanel.Dock="Right">
                         <Path Data="{StaticResource SettingsRegular}" Fill="{DynamicResource TextControlForeground}" />
+                        <Button.Flyout>
+                            <Flyout>
+                                <StackPanel Orientation="Vertical" Spacing="5">
+                                    <TextBlock Text="{x:Static lang:Lang.Settings_Theme}" />
+                                    <StackPanel Orientation="Vertical" Spacing="5">
+                                        <RadioButton Name="DefaultTheme"
+                                                     Content="{x:Static lang:Lang.Settings_ThemeDefault}"
+                                                     IsCheckedChanged="SystemThemeChecked" />
+                                        <RadioButton Name="LightTheme"
+                                                     Content="{x:Static lang:Lang.Settings_ThemeLight}"
+                                                     IsCheckedChanged="LightThemeChecked" />
+                                        <RadioButton Name="DarkTheme" Content="{x:Static lang:Lang.Settings_ThemeDark}"
+                                                     IsCheckedChanged="DarkThemeChecked" />
+                                    </StackPanel>
+                                    <ToggleSwitch OnContent="{x:Static lang:Lang.Settings_UseBlur}"
+                                                  IsVisible="{CompiledBinding $parent[views:MainView].IsOnDesktop}"
+                                                  IsEnabled="{CompiledBinding $self.IsVisible}" Name="UseBlur"
+                                                  OffContent="{CompiledBinding $self.OnContent}"
+                                                  IsChecked="{CompiledBinding $parent[views:MainView].Settings.UseBlur, Mode=TwoWay}"
+                                                  IsCheckedChanged="UseBlurCheckedChanged" />
+                                    <DockPanel IsVisible="{CompiledBinding #UseBlur.IsChecked}"
+                                               IsEnabled="{CompiledBinding #UseBlur.IsChecked}">
+                                        <TextBlock Text="{x:Static lang:Lang.Settings_BlurLevel}"
+                                                   VerticalAlignment="Center"
+                                                   DockPanel.Dock="Left" />
+                                        <TextBlock Text="%" DockPanel.Dock="Right" VerticalAlignment="Center" />
+                                        <TextBlock Text="{CompiledBinding #BlurLevel.Value}" DockPanel.Dock="Right"
+                                                   VerticalAlignment="Center" />
+                                        <Slider Value="{CompiledBinding $parent[views:MainView].Settings.BlurLevel}"
+                                                TickFrequency="1" MinWidth="100"
+                                                Margin="5 0"
+                                                Name="BlurLevel" IsSnapToTickEnabled="True"
+                                                ValueChanged="BlurLevelValueChanged" />
+                                    </DockPanel>
+                                </StackPanel>
+                            </Flyout>
+                        </Button.Flyout>
                     </Button>
                     <Button HotKey="Ctrl+N" Name="NewNoteButton" HorizontalAlignment="Stretch"
                             HorizontalContentAlignment="Center"
@@ -276,11 +312,6 @@
                            Text="{x:Static lang:Lang.MainView_Welcome}" />
             </DockPanel>
             <views:AboutScreen IsEnabled="False" Name="AboutView" Main="{CompiledBinding $parent[views:MainView]}" />
-            <views:SettingsScreen IsEnabled="False" Name="SettingsView"
-                                  IsOnDesktop="{CompiledBinding $parent[views:MainView].IsOnDesktop}"
-                                  DesktopContainer="{CompiledBinding $parent[views:MainView].DesktopContainer}"
-                                  Settings="{CompiledBinding $parent[views:MainView].Settings}"
-                                  Main="{CompiledBinding $parent[views:MainView]}" />-
         </Carousel>
     </dialogHost:DialogHost>
 </UserControl>
