+++ src/Mindbank/Views/MainView.axaml.cs
@@ -6,11 +6,13 @@ using System.Reflection;
 using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Controls;
+using Avalonia.Controls.Primitives;
 using Avalonia.Controls.Shapes;
 using Avalonia.Interactivity;
 using Avalonia.Layout;
 using Avalonia.Media;
 using Avalonia.Platform.Storage;
+using Avalonia.Styling;
 using Avalonia.Threading;
 using DialogHostAvalonia;
 using Mindbank.Backend;
@@ -36,6 +38,8 @@ public partial class MainView : UserControl
 
     internal static readonly StyledProperty<bool> UpdateAvailableProperty =
         AvaloniaProperty.Register<MainView, bool>(nameof(UpdateAvailable));
+
+    private bool _initializingSettings = true;
 
     public MainView()
     {
@@ -197,7 +201,14 @@ public partial class MainView : UserControl
 
     private void Init(object? sender, EventArgs e)
     {
+        _initializingSettings = true;
         Settings.Load();
+        if (Settings.Theme == ThemeVariant.Default && DefaultTheme != null) DefaultTheme.IsChecked = true;
+        if (Settings.Theme == ThemeVariant.Dark && DarkTheme != null) DarkTheme.IsChecked = true;
+        if (Settings.Theme == ThemeVariant.Light && LightTheme != null) LightTheme.IsChecked = true;
+        if (UseBlur != null) UseBlur.IsChecked = Settings.UseBlur;
+        if (BlurLevel != null) BlurLevel.Value = Settings.BlurLevel;
+        _initializingSettings = false;
     }
 
 
@@ -286,7 +297,7 @@ public partial class MainView : UserControl
         var l = new List<object?>();
         foreach (var control in MainCarousel.Items)
         {
-            if (Equals(control, MainPanel) || Equals(control, SettingsView) || Equals(control, AboutView)) continue;
+            if (Equals(control, MainPanel) || Equals(control, AboutView)) continue;
             l.Add(control);
             if (control is NoteScreen ns) ns.Main = null;
         }
@@ -388,5 +399,54 @@ public partial class MainView : UserControl
     private void CloseDialogHost(object? s, RoutedEventArgs e)
     {
         MainDialogHost.CurrentSession?.Close();
+    }
+
+    private void SystemThemeChecked(object? sender, RoutedEventArgs e)
+    {
+        if (_initializingSettings || sender is not RadioButton { IsChecked: true } ||
+            Application.Current is not { } app) return;
+        app.RequestedThemeVariant = ThemeVariant.Default;
+        Settings.Theme = ThemeVariant.Default;
+        if (BlurLevel is { Value: var v } && DesktopContainer is not null && UseBlur is { IsChecked: var useBlur })
+            DesktopContainer.SetOpacity(useBlur is true ? v : 100);
+        if (!Design.IsDesignMode) Settings.Save();
+    }
+
+    private void LightThemeChecked(object? sender, RoutedEventArgs e)
+    {
+        if (_initializingSettings || sender is not RadioButton { IsChecked: true } ||
+            Application.Current is not { } app) return;
+        app.RequestedThemeVariant = ThemeVariant.Light;
+        Settings.Theme = ThemeVariant.Light;
+        if (BlurLevel is { Value: var v } && DesktopContainer is not null && UseBlur is { IsChecked: var useBlur })
+            DesktopContainer.SetOpacity(useBlur is true ? v : 100);
+        if (!Design.IsDesignMode) Settings.Save();
+    }
+
+    private void DarkThemeChecked(object? sender, RoutedEventArgs e)
+    {
+        if (_initializingSettings || sender is not RadioButton { IsChecked: true } ||
+            Application.Current is not { } app) return;
+        app.RequestedThemeVariant = ThemeVariant.Dark;
+        Settings.Theme = ThemeVariant.Dark;
+        if (BlurLevel is { Value: var v } && DesktopContainer is not null && UseBlur is { IsChecked: var useBlur })
+            DesktopContainer.SetOpacity(useBlur is true ? v : 100);
+        if (!Design.IsDesignMode) Settings.Save();
+    }
+
+    private void BlurLevelValueChanged(object? sender, RangeBaseValueChangedEventArgs e)
+    {
+        if (_initializingSettings || sender is not Slider { Value: var v } || DesktopContainer is null ||
+            UseBlur is not { IsChecked: var useBlur }) return;
+        DesktopContainer.SetOpacity(useBlur is true ? v : 100);
+        if (!Design.IsDesignMode) Settings.Save();
+    }
+
+    private void UseBlurCheckedChanged(object? sender, RoutedEventArgs e)
+    {
+        if (_initializingSettings || BlurLevel is not { Value: var v } || DesktopContainer is null ||
+            UseBlur is not { IsChecked: var useBlur }) return;
+        DesktopContainer.SetOpacity(useBlur is true ? v : 100);
+        if (!Design.IsDesignMode) Settings.Save();
     }
 }
